Tratamento de erros e validações

+++ src/main/java/com/apilivraria/service/AutorService.java
@@ -1,7 +1,11 @@
 package com.apilivraria.service;
 
+import com.apilivraria.exceptions.OperacaoNaoPermitidaException;
 import com.apilivraria.model.Autor;
 import com.apilivraria.repository.AutorRepository;
+import com.apilivraria.repository.LivroRepository;
+import com.apilivraria.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,15 +13,15 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository autorRepository;
-
-    public AutorService(AutorRepository autorRepository) {
-        this.autorRepository = autorRepository;
-    }
+    private final AutorValidator autorValidator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
+        autorValidator.validar(autor);
         return autorRepository.save(autor);
     }
 
@@ -25,6 +29,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("ID não encontrado");
         }
+        autorValidator.validar(autor);
         autorRepository.save(autor);
     }
 
@@ -33,6 +38,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não e permitido excluir um Autor que possui Livros cadastrados");
+        }
         autorRepository.delete(autor);
     }
 
@@ -50,5 +58,9 @@ public class AutorService {
         }
 
         return autorRepository.findAll();
+    }
+
+    private boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
